+++ extras/OneFileCMS.config.SAMPLE.php
@@ -1,6 +1,6 @@
 <?php 
 // OneFileCMS - github.com/Self-Evident/OneFileCMS
-// v3.5.04+
+// v3.5.05
 // Sample external config file - this file is OPTIONAL.
 //
 // Basically, what follows is just a copy & paste of the OneFileCMS CONFIGURABLE INFO section.
@@ -17,7 +17,8 @@ $SALT     = 'somerandomsalt';
 $MAX_ATTEMPTS  = 3;   //Max failed login attempts before LOGIN_DELAY starts.
 $LOGIN_DELAY   = 10;  //In seconds.
 $MAX_IDLE_TIME = 600; //In seconds. 600 = 10 minutes.  Other PHP settings (like gc) may limit its max effective value.
-	$MAX_IDLE_TIME = 60000; //In seconds. 600 = 10 minutes.  Other PHP settings (like gc) may limit its max effective value. //##### 
+$TO_WARNING    = 120; //In seconds. When idle time remaining is less than this value, a timeout warning is displayed.
+
 $MAIN_WIDTH    = '810px'; //Width of main <div> defining page layout.          Can be px, pt, em, or %.  Assumes px otherwise.
 $WIDE_VIEW_WIDTH = '97%'; //Width to set Edit page if [Wide View] is clicked.  Can be px, pt, em, or %.  Assumes px otherwise.
 
@@ -55,8 +56,8 @@ $SESSION_NAME = 'OFCMS'; //Name of session cookie. Change if using multiple copi
 //$ACCESS_ROOT = 'some/path/';
 
 
-//URL of optional external style sheet.  Used as href in <link ...>
-//If file is not found, or is incomplete, the built-in defaults will be used.
+//URL of optional external style sheet.  Used as an href in <link ...>
+//If file is not found, or is incomplete, built-in defaults will be used.
 //$CSS_FILE = 'OneFileCMS.css';